stripe

+++ app/api/board/route.js
@@ -24,6 +24,15 @@ export async function POST(req) {
     await connectMongo();
     const user = await User.findById(session.user.id);
 
+    if (!user.hasAccess) {
+      return NextResponse.json(
+        {
+          error: "Please subscribe first",
+        },
+        { status: 403 }
+      );
+    }
+
     const board = await Board.create({
       userId: user._id,
       name: body.name,
@@ -40,10 +49,8 @@ export async function POST(req) {
 
 export async function DELETE(req) {
   try {
-    //const searchParams = req.nextUrl.searchParams
     const { searchParams } = req.nextUrl;
     const boardId = searchParams.get("boardId");
-    const session = await auth();
 
     if (!boardId) {
       return NextResponse.json(
@@ -52,12 +59,26 @@ export async function DELETE(req) {
       );
     }
 
+    const session = await auth();
+
+    if (!session) {
+      return NextResponse.json({ error: "Not authorized" }, { status: 401 });
+    }
+
+    const user = await User.findById(session?.user?.id);
+
+    if (!user.hasAccess) {
+      return NextResponse.json(
+        { error: "Please subscribe first" },
+        { status: 403 }
+      );
+    }
+
     await Board.deleteOne({
       _id: boardId,
       userId: session?.user?.id,
     });
 
-    const user = await User.findById(session?.user?.id);
     user.boards = user.boards.filter((id) => id.toString() !== boardId);
     await user.save();
